+++ comb_spec_searcher/strategies/constructor/base.py
@@ -26,6 +26,13 @@ from comb_spec_searcher.typing import (
 class Constructor(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectType]):
     """The constructor is akin to the 'counting function' in the comb exp paper."""
 
+    def can_be_equivalent(self) -> bool:
+        """
+        Return False if the constructor can NOT be an 'identity' map
+        (i.e., up to rearranging params)
+        """
+        return True
+
     @abc.abstractmethod
     def get_equation(
         self, lhs_func: sympy.Function, rhs_funcs: Tuple[sympy.Function, ...]